Reimplement bazel-latex to use its own copy of TeXLive.
Instead of depending on the host system to provide a full TeXLive
installation, this package now downloads a copy instead. We're not using
the official upstream one, as it is gigantic in size. A separate
GitHub repository has been created, hosting a decomposed version:

	https://github.com/ProdriveTechnologies/texlive-modular

By default, we depend on a compact installation of TeXLive. If a user is
interested in custom fonts or packages, they need to be added to the
sources manually. The documentation will be updated to reflect that.