Invoke pdflatex through latexrun.
latexrun is a script that tidies up the output of pdflatex in such a way
that it looks similar to GCC/Clang. (Almost?) all unnecessary output is
removed. Instead of suppressing the output on success, now use the
occasion to simply print the output of this command, so the user has
useful warnings to act on.

Unfortunately, the stock version of latexrun required a minor change to
work properly. It seems to crash by default if pdflatex invokes mktexfmt
as well. Fix this by using non-greedy pattern matching. Also improve the
output a bit by using relative paths.

Fixes: #3